Move memory and extensions surface tests
