+++ Tests/QuillCodeAppTests/WorkspaceProjectExtensionIntegrationTests.swift
@@ -1,4 +1,5 @@
 import XCTest
+import QuillCodeCore
 @testable import QuillCodeApp
 
 @MainActor
@@ -17,6 +18,59 @@ final class WorkspaceProjectExtensionIntegrationTests: XCTestCase {
         XCTAssertEqual(extensions.mcpServerCount, 0)
         XCTAssertEqual(extensions.items.first?.name, "GitHub")
         XCTAssertEqual(extensions.items.first?.relativePath, ".quillcode/plugins/github.json")
+    }
+
+    func testSurfaceIncludesProjectExtensionSummaryAndCommand() {
+        let project = ProjectRef(
+            name: "QuillCode",
+            path: "/tmp/QuillCode",
+            extensionManifests: [
+                ProjectExtensionManifest(
+                    id: "plugin:github",
+                    kind: .plugin,
+                    name: "GitHub",
+                    summary: "PR workflow helpers.",
+                    version: "1.2.0",
+                    sourceURL: "https://github.com/Lore-Hex/quillcode-github",
+                    relativePath: ".quillcode/plugins/github.json",
+                    updateCommand: "git -C .quillcode/plugins/github pull --ff-only",
+                    updateTimeoutSeconds: 300
+                ),
+                ProjectExtensionManifest(
+                    id: "mcp_server:filesystem",
+                    kind: .mcpServer,
+                    name: "Filesystem MCP",
+                    summary: "Workspace MCP server.",
+                    relativePath: ".quillcode/mcp/filesystem.json",
+                    transport: .stdio,
+                    launchExecutable: "quill-mcp",
+                    launchCommand: "quill-mcp --root .",
+                    launchArguments: ["--root", "."]
+                )
+            ]
+        )
+        let model = QuillCodeWorkspaceModel(
+            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
+            extensions: ExtensionsState(isVisible: true)
+        )
+
+        let surface = model.surface()
+
+        XCTAssertEqual(surface.extensions.subtitle, "1 plugin · 0 skills · 1 MCP server")
+        XCTAssertEqual(surface.extensions.items.map(\.kindLabel), ["Plugin", "MCP"])
+        XCTAssertEqual(surface.extensions.items.map(\.statusLabel), ["Discovered", "Stopped"])
+        XCTAssertEqual(surface.extensions.items.first?.versionLabel, "v1.2.0")
+        XCTAssertEqual(surface.extensions.items.first?.sourceURL, "https://github.com/Lore-Hex/quillcode-github")
+        XCTAssertEqual(surface.extensions.items.first?.updateCommandID, "extension-update:plugin:github")
+        XCTAssertEqual(surface.commands.first { $0.id == "extension-update:plugin:github" }?.title, "Update GitHub")
+        XCTAssertEqual(surface.commands.first { $0.id == "extension-update:plugin:github" }?.isEnabled, true)
+        XCTAssertEqual(surface.extensions.items.last?.transportLabel, "STDIO")
+        XCTAssertEqual(surface.extensions.items.last?.launchCommand, "quill-mcp --root .")
+        XCTAssertEqual(surface.extensions.items.last?.startCommandID, "mcp-start:mcp_server:filesystem")
+        XCTAssertEqual(surface.commands.first { $0.id == "mcp-start:mcp_server:filesystem" }?.isEnabled, true)
+        XCTAssertEqual(surface.commands.first { $0.id == "mcp-stop:mcp_server:filesystem" }?.isEnabled, false)
+        XCTAssertEqual(surface.commands.first { $0.id == "toggle-extensions" }?.category, WorkspaceCommandPalette.extensionsCategory)
+        XCTAssertEqual(surface.commands.first { $0.id == "toggle-extensions" }?.isEnabled, true)
     }
 
     func testProjectExtensionUpdateCommandRunsAndRefreshesProjectMetadata() throws {
