+++ Tests/QuillCodeAppTests/WorkspaceSurfaceTests.swift
@@ -1,6 +1,5 @@
 import XCTest
 import QuillCodeCore
-import QuillCodeTools
 @testable import QuillCodeApp
 
 @MainActor
@@ -486,173 +485,6 @@ final class WorkspaceSurfaceTests: XCTestCase {
         XCTAssertTrue(command?.keywords.contains("QUILL_ENV") == true)
         XCTAssertTrue(command?.keywords.contains("app") == true)
         XCTAssertTrue(command?.keywords.contains("120s") == true)
-    }
-
-    func testSurfaceIncludesProjectExtensionSummaryAndCommand() {
-        let project = ProjectRef(
-            name: "QuillCode",
-            path: "/tmp/QuillCode",
-            extensionManifests: [
-                ProjectExtensionManifest(
-                    id: "plugin:github",
-                    kind: .plugin,
-                    name: "GitHub",
-                    summary: "PR workflow helpers.",
-                    version: "1.2.0",
-                    sourceURL: "https://github.com/Lore-Hex/quillcode-github",
-                    relativePath: ".quillcode/plugins/github.json",
-                    updateCommand: "git -C .quillcode/plugins/github pull --ff-only",
-                    updateTimeoutSeconds: 300
-                ),
-                ProjectExtensionManifest(
-                    id: "mcp_server:filesystem",
-                    kind: .mcpServer,
-                    name: "Filesystem MCP",
-                    summary: "Workspace MCP server.",
-                    relativePath: ".quillcode/mcp/filesystem.json",
-                    transport: .stdio,
-                    launchExecutable: "quill-mcp",
-                    launchCommand: "quill-mcp --root .",
-                    launchArguments: ["--root", "."]
-                )
-            ]
-        )
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
-            extensions: ExtensionsState(isVisible: true)
-        )
-
-        let surface = model.surface()
-
-        XCTAssertEqual(surface.extensions.subtitle, "1 plugin · 0 skills · 1 MCP server")
-        XCTAssertEqual(surface.extensions.items.map(\.kindLabel), ["Plugin", "MCP"])
-        XCTAssertEqual(surface.extensions.items.map(\.statusLabel), ["Discovered", "Stopped"])
-        XCTAssertEqual(surface.extensions.items.first?.versionLabel, "v1.2.0")
-        XCTAssertEqual(surface.extensions.items.first?.sourceURL, "https://github.com/Lore-Hex/quillcode-github")
-        XCTAssertEqual(surface.extensions.items.first?.updateCommandID, "extension-update:plugin:github")
-        XCTAssertEqual(surface.commands.first { $0.id == "extension-update:plugin:github" }?.title, "Update GitHub")
-        XCTAssertEqual(surface.commands.first { $0.id == "extension-update:plugin:github" }?.isEnabled, true)
-        XCTAssertEqual(surface.extensions.items.last?.transportLabel, "STDIO")
-        XCTAssertEqual(surface.extensions.items.last?.launchCommand, "quill-mcp --root .")
-        XCTAssertEqual(surface.extensions.items.last?.startCommandID, "mcp-start:mcp_server:filesystem")
-        XCTAssertEqual(surface.commands.first { $0.id == "mcp-start:mcp_server:filesystem" }?.isEnabled, true)
-        XCTAssertEqual(surface.commands.first { $0.id == "mcp-stop:mcp_server:filesystem" }?.isEnabled, false)
-        XCTAssertEqual(surface.commands.first { $0.id == "toggle-extensions" }?.category, WorkspaceCommandPalette.extensionsCategory)
-        XCTAssertEqual(surface.commands.first { $0.id == "toggle-extensions" }?.isEnabled, true)
-    }
-
-    func testSurfaceShowsReadyMCPServerProbeSummaryAndStopAction() {
-        let project = ProjectRef(
-            name: "QuillCode",
-            path: "/tmp/QuillCode",
-            extensionManifests: [
-                ProjectExtensionManifest(
-                    id: "mcp_server:filesystem",
-                    kind: .mcpServer,
-                    name: "Filesystem MCP",
-                    relativePath: ".quillcode/mcp/filesystem.json",
-                    transport: .stdio,
-                    launchExecutable: "quill-mcp",
-                    launchCommand: "quill-mcp --root .",
-                    launchArguments: ["--root", "."]
-                )
-            ]
-        )
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
-            extensions: ExtensionsState(
-                isVisible: true,
-                mcpServerStatuses: ["mcp_server:filesystem": .ready],
-                mcpServerProbeSummaries: [
-                    "mcp_server:filesystem": MCPServerProbeSummary(
-                        protocolVersion: "2024-11-05",
-                        serverName: "Fixture MCP",
-                        serverVersion: "1.0.0",
-                        toolDescriptors: [
-                            MCPToolDescriptor(
-                                name: "read_file",
-                                description: "Read a file",
-                                requiredArguments: ["path"],
-                                schemaSummary: "required: path:string"
-                            ),
-                            MCPToolDescriptor(
-                                name: "write_file",
-                                requiredArguments: ["content", "path"],
-                                optionalArguments: ["overwrite"],
-                                schemaSummary: "required: content:string, path:string; optional: overwrite:boolean"
-                            )
-                        ],
-                        resourceNames: ["README", "Project config"],
-                        promptNames: ["summarize_project"]
-                    )
-                ]
-            )
-        )
-
-        let surface = model.surface()
-
-        XCTAssertEqual(surface.extensions.items.first?.statusLabel, "Ready")
-        XCTAssertEqual(surface.extensions.items.first?.serverLabel, "Fixture MCP 1.0.0")
-        XCTAssertEqual(surface.extensions.items.first?.protocolLabel, "MCP 2024-11-05")
-        XCTAssertEqual(surface.extensions.items.first?.toolCountLabel, "2 tools")
-        XCTAssertEqual(surface.extensions.items.first?.toolNames, ["read_file", "write_file"])
-        XCTAssertEqual(surface.extensions.items.first?.toolDescriptors.map(\.schemaSummary), [
-            "required: path:string",
-            "required: content:string, path:string; optional: overwrite:boolean"
-        ])
-        XCTAssertEqual(surface.extensions.items.first?.resourceCountLabel, "2 resources")
-        XCTAssertEqual(surface.extensions.items.first?.resourceNames, ["README", "Project config"])
-        XCTAssertEqual(surface.extensions.items.first?.promptCountLabel, "1 prompt")
-        XCTAssertEqual(surface.extensions.items.first?.promptNames, ["summarize_project"])
-        XCTAssertNil(surface.extensions.items.first?.startCommandID)
-        XCTAssertEqual(surface.extensions.items.first?.stopCommandID, "mcp-stop:mcp_server:filesystem")
-        XCTAssertEqual(surface.commands.first { $0.id == "mcp-start:mcp_server:filesystem" }?.isEnabled, false)
-        XCTAssertEqual(surface.commands.first { $0.id == "mcp-stop:mcp_server:filesystem" }?.isEnabled, true)
-        XCTAssertEqual(surface.commands.first { $0.id == "stop-all" }?.isEnabled, true)
-    }
-
-    func testSurfaceIncludesMemorySummariesAndCommand() {
-        let project = ProjectRef(
-            name: "QuillCode",
-            path: "/tmp/QuillCode",
-            memories: [
-                MemoryNote(
-                    id: "project:.quillcode/memories/project.md",
-                    scope: .project,
-                    title: "Project",
-                    content: "QuillCode should stay native Swift and document major decisions.",
-                    relativePath: ".quillcode/memories/project.md",
-                    byteCount: 63
-                )
-            ]
-        )
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(
-                projects: [project],
-                selectedProjectID: project.id,
-                globalMemories: [
-                    MemoryNote(
-                        id: "global:memories/preferences.md",
-                        scope: .global,
-                        title: "Preferences",
-                        content: "Prefer focused tests and small reviewable commits.",
-                        relativePath: "memories/preferences.md",
-                        byteCount: 48
-                    )
-                ]
-            ),
-            memories: MemoriesState(isVisible: true)
-        )
-
-        let surface = model.surface()
-
-        XCTAssertTrue(surface.memories.isVisible)
-        XCTAssertEqual(surface.memories.globalCount, 1)
-        XCTAssertEqual(surface.memories.projectCount, 1)
-        XCTAssertEqual(surface.memories.items.map { $0.scope }, [MemoryScope.global, .project])
-        XCTAssertEqual(surface.memories.items.first?.title, "Preferences")
-        XCTAssertEqual(surface.topBar.memoryLabel, "2 memories")
-        XCTAssertEqual(surface.commands.first { $0.id == "toggle-memories" }?.category, WorkspaceCommandPalette.memoriesCategory)
     }
 
     func testStopAllCommandIsEnabledForTerminalRuns() {
